requirements.txt
simple-salesforce==1.12.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
